+++ src/router/index.ts
@@ -60,4 +60,7 @@ const routes: Array<RouteConfig> = [
 export default new VueRouter({
   mode: "history",
   routes,
+  scrollBehavior() {
+    return { x: 0, y: 0 };
+  },
 });